Add custom logging to `seqio.evaluation.Evaluator`.
This adds the ability to pass a function to the Evaluator via the `log_fn`
parameter. When running evaluation this function with the resulting task
metrics. If a `log_fn` isn't provided it will default to using the old
`_log_eval_results` method.

This also includes some updates the the evaluation tests so the default
`_log_fn` is set when we mock the `Evaluator`'s `__init__` and adds a test
for the newly defined custom logging function in p5x.

PiperOrigin-RevId: 353045215